Removed test/dloop/DLoopCoreMock/calculation-test.ts and update DLoopCoreLogic test to increase the coverage
