+++ typescript/common/bps_constants.ts
@@ -2,3 +2,4 @@
 export const ONE_BPS_UNIT = 100; // 1 bps with 2 decimals
 export const ONE_PERCENT_BPS = 100 * ONE_BPS_UNIT; // 1% in basis points
 export const ONE_HUNDRED_PERCENT_BPS = 100 * ONE_PERCENT_BPS; // 100% in basis points
+export const TARGET_LEVERAGE_BPS = 300 * ONE_PERCENT_BPS; // 3x leverage (300%)
